Encapsulate state maps in the context interface.
Been meaning to do this for quite a while
