+++ choam/src/main/java/com/salesforce/apollo/choam/GenesisAssembly.java
@@ -32,7 +32,6 @@ import com.salesfoce.apollo.choam.proto.Validations;
 import com.salesfoce.apollo.choam.proto.ViewMember;
 import com.salesfoce.apollo.utils.proto.PubKey;
 import com.salesforce.apollo.choam.comm.Terminal;
-import com.salesforce.apollo.choam.fsm.BrickLayer;
 import com.salesforce.apollo.choam.fsm.Genesis;
 import com.salesforce.apollo.choam.support.HashedBlock;
 import com.salesforce.apollo.choam.support.HashedCertifiedBlock;
