+++ choam/src/main/java/com/salesforce/apollo/choam/Producer.java
@@ -36,8 +36,8 @@ import com.salesfoce.apollo.choam.proto.UnitData;
 import com.salesfoce.apollo.choam.proto.Validate;
 import com.salesforce.apollo.choam.comm.Terminal;
 import com.salesforce.apollo.choam.fsm.Driven;
+import com.salesforce.apollo.choam.fsm.Driven.Earner;
 import com.salesforce.apollo.choam.fsm.Driven.Transitions;
-import com.salesforce.apollo.choam.fsm.Earner;
 import com.salesforce.apollo.choam.support.HashedBlock;
 import com.salesforce.apollo.choam.support.HashedCertifiedBlock;
 import com.salesforce.apollo.choam.support.TxDataSource;
