+++ choam/src/main/java/com/salesforce/apollo/choam/fsm/Genesis.java
@@ -6,6 +6,7 @@
  */
 package com.salesforce.apollo.choam.fsm;
 
+import com.chiralbehaviors.tron.Entry;
 import com.chiralbehaviors.tron.FsmExecutor;
 import com.salesforce.apollo.ethereal.Ethereal.PreBlock;
 
@@ -14,7 +15,67 @@ import com.salesforce.apollo.ethereal.Ethereal.PreBlock;
  *
  */
 public interface Genesis {
-    public interface Transitions extends FsmExecutor<Genesis, Genesis.Transitions> {
+    enum BrickLayer implements Transitions {
+
+        CERTIFICATION {
+            @Entry
+            public void certify() {
+                context().certify();
+            }
+
+            @Override
+            public Transitions process(PreBlock preblock, boolean last) {
+                context().certify(preblock, last);
+                return last ? PUBLISH : null;
+            }
+        },
+        FAIL {
+        },
+        INITIAL {
+            @Entry
+            public void gather() {
+                context().gather();
+            }
+
+            @Override
+            public Transitions nextEpoch(Integer epoch) {
+                return epoch.equals(0) ? null : NOMINATION;
+
+            }
+
+            @Override
+            public Transitions process(PreBlock preblock, boolean last) {
+                context().gather(preblock, last);
+                return null;
+            }
+        },
+        NOMINATION {
+            @Override
+            public Transitions nextEpoch(Integer epoch) {
+                return CERTIFICATION;
+            }
+
+            @Entry
+            public void nominate() {
+                context().nominate();
+            }
+
+            @Override
+            public Transitions process(PreBlock preblock, boolean last) {
+                context().nominations(preblock, last);
+                return null;
+            }
+        },
+        PUBLISH {
+            @Entry
+            public void publish() {
+                context().publish();
+            }
+        };
+
+    }
+
+    interface Transitions extends FsmExecutor<Genesis, Genesis.Transitions> {
 
         default Transitions nextEpoch(Integer epoch) {
             throw fsm().invalidTransitionOn();
